+++ Sources/DeckKit/_Deprecated/Favorites/FavoriteStore.swift
@@ -8,8 +8,7 @@
 
 import Foundation
 
-/// This protocol can be implemented by any type that can be
-/// used to store and manage favorites.
+@available(*, deprecated, message: "This class has been moved to https://github.com/danielsaidi/SwiftUIKit")
 public protocol FavoriteStore: AnyObject {
 
     /// The item type that is managed by the store.